Rename TranslationFirstDualQuaternion
