+++ s2e-ff/src/library/math/translation_first_dual_quaternion.hpp
