+++ s2e-ff/test/TestTranslationFirstDualQuaternion.cpp
@@ -1,6 +1,6 @@
 #include <gtest/gtest.h>
 
-#include "../src/Library/math/TranslationFirstDualQuaternion.hpp"
+#include "../src/Library/math/translation_first_dual_quaternion.hpp"
 
 TEST(TranslationFirstDualQuaternion, ConstructorFromRotationTranslation) {
   libra::Quaternion q_rot(1.0, 0.0, 0.0, 1.0);  // 90deg rotation around X axis
